fix(text): properly handle utf-8 chars
UTF-8 characters are now handled porperly. Also re-worked some helper
functions with UTF-8 in mind.

I secretly no real C developers will look at this. I might have to ask
for witness protection.

Closes #43

+++ include/swappy.h
@@ -41,7 +41,7 @@ struct swappy_paint_text {
   double s;
   gchar *font;
   gchar *text;
-  size_t cursor;
+  glong cursor;
   struct swappy_point from;
   struct swappy_point to;
   enum swappy_text_mode mode;
